feat(registry-api): rewriting serverless function
